+++ src/feelpp/benchmarking/report/atomicReports/controller.py
@@ -1,0 +1,4 @@
+from feelpp.benchmarking.report.base.controller import Controller
+
+class AtomicReportController(Controller):
+    pass
