+++ src/feelpp/benchmarking/report/atomicReports/view.py
@@ -1,0 +1,4 @@
+from feelpp.benchmarking.report.base.view import View
+
+class AtomicReportView(View):
+    pass
