+++ src/feelpp/benchmarking/report/base/model.py
@@ -1,0 +1,7 @@
+class Model:
+    """ Abstract model component """
+    def __init__(self):
+        pass
+
+    def buildMasterDf(self):
+        pass
