+++ src/feelpp/benchmarking/report/overview/controller.py
@@ -1,0 +1,4 @@
+from feelpp.benchmarking.report.base.controller import Controller
+
+class OverviewController(Controller):
+    pass
