+++ src/feelpp/benchmarking/report/strategies.py
